requirements.txt
pandas==2.2.3
arcgis==2.4.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
geoalchemy2==0.15.2